move entities, add id fields
