+++ Core/Entities/HubCommunicator.cs
@@ -1,4 +1,4 @@
-﻿namespace Core;
+﻿namespace Core.Entities;
 
 public class HubCommunicator
 {
