+++ Core/Entities/Vendor.cs
@@ -1,20 +1,15 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Core
+﻿namespace Core.Entities
 {
     internal class Vendor
     {
+        public int Id { get; set; }
         private string _name { get; }
         private Coordinates _coordinates { get; }
 
         public Vendor(string name, Coordinates coordinates)
         {
-            this._name = name;
-            this._coordinates = coordinates;
+            _name = name;
+            _coordinates = coordinates;
         }
     }
 }
