--- Core/Product.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Core
-{
-    internal class Product
-    {
-        private string _name { get; }
-
-        public Product(string name)
-        {
-            this._name = name;
-        }
-    }
-}
